refactor!: hide default_settings
