+++ expertsystem/ui/__init__.py
@@ -1,9 +1,7 @@
 """User Interface of the `expertsystem`."""
 
 __all__ = [
-    "default_settings",
     "system_control",
 ]
 
-from . import default_settings
 from . import system_control
