+++ tests/canonical_formalism/test_ls_coupling.py
@@ -7,7 +7,7 @@ from expertsystem.state.particle import (
     Spin,
     _SpinQNConverter,
 )
-from expertsystem.ui.default_settings import (
+from expertsystem.ui._default_settings import (
     create_default_interaction_settings,
 )
 from expertsystem.ui.system_control import (
