+++ tox.ini
@@ -128,7 +128,7 @@ ignore_errors = True
 ignore_errors = False
 [mypy-expertsystem.io.*]
 ignore_errors = False
-[mypy-expertsystem.ui.default_settings.*]
+[mypy-expertsystem.ui._default_settings.*]
 ignore_errors = False
 
 [pydocstyle]
